refactor: moved QueryIdCachingProxyHandler to Constants.java
